Renamed aikido::path -> ::trajectory.
