+++ aikido/include/aikido/planner/SnapPlanner.hpp
@@ -3,7 +3,7 @@
 #include "../statespace/StateSpace.hpp"
 #include "../statespace/Interpolator.hpp"
 #include "../constraint/TestableConstraint.hpp"
-#include "../path/Interpolated.hpp"
+#include "../trajectory/Interpolated.hpp"
 #include "PlanningResult.hpp"
 
 namespace aikido {
@@ -22,7 +22,7 @@ namespace planner {
 /// \param constraint trajectory-wide constraint that must be satisfied
 /// \param[out] planningResult information about success or failure
 /// \return trajectory or \c nullptr if planning failed
-path::InterpolatedPtr planSnap(
+trajectory::InterpolatedPtr planSnap(
   const std::shared_ptr<statespace::StateSpace>& stateSpace,
   const statespace::StateSpace::State *startState,
   const statespace::StateSpace::State *goalState,
