+++ aikido/include/aikido/trajectory/Spline.hpp
@@ -1,9 +1,9 @@
-#ifndef AIKIDO_PATH_SPLINETRAJECTORY2_HPP_
-#define AIKIDO_PATH_SPLINETRAJECTORY2_HPP_
+#ifndef AIKIDO_TRAJECTORY_SPLINETRAJECTORY2_HPP_
+#define AIKIDO_TRAJECTORY_SPLINETRAJECTORY2_HPP_
 #include "Trajectory.hpp"
 
 namespace aikido {
-namespace path {
+namespace trajectory {
 
 /// Polynomial spline trajectory defined in a \c StateSpace. The trajectory is
 /// represented as a sequence of polynomial segments between knot points. Each
@@ -116,7 +116,7 @@ private:
   std::vector<PolynomialSegment> mSegments;
 };
 
-} // namespace path
+} // namespace trajectory
 } // namespace aikido
 
-#endif // ifndef AIKIDO_PATH_SPLINETRAJECTORY2_HPP_
+#endif // ifndef AIKIDO_TRAJECTORY_SPLINETRAJECTORY2_HPP_
