+++ aikido/include/aikido/trajectory/detail/SplineTrajectory-impl.hpp
