+++ aikido/src/CMakeLists.txt
@@ -2,8 +2,8 @@ add_subdirectory("external/hauser_parabolic_smoother")
 
 add_subdirectory("util")
 add_subdirectory("statespace")
-add_subdirectory("path")       # Dependencies: statespace
-add_subdirectory("constraint") # Dependencies: statespace, util
-add_subdirectory("planner")    # Dependencies: external, statespace, path
-add_subdirectory("ompl")       # Dependencies: constraint, statespace, path
 add_subdirectory("distance")   # Dependencies: statespace
+add_subdirectory("trajectory") # Dependencies: statespace
+add_subdirectory("constraint") # Dependencies: statespace, util
+add_subdirectory("planner")    # Dependencies: external, statespace, trajectory
+add_subdirectory("ompl")       # Dependencies: constraint, statespace, trajectory
