+++ aikido/src/ompl/CMakeLists.txt
@@ -9,7 +9,7 @@ target_link_libraries("${PROJECT_NAME}_ompl"
   "${PROJECT_NAME}_constraint"
   "${PROJECT_NAME}_distance"
   "${PROJECT_NAME}_statespace"
-  "${PROJECT_NAME}_path"
+  "${PROJECT_NAME}_trajectory"
   ${Boost_LIBRARIES}
   ${DART_LIBRARIES}
   ${OMPL_LIBRARIES}
