+++ aikido/src/planner/CMakeLists.txt
@@ -4,6 +4,6 @@ aikido_add_library("${PROJECT_NAME}_planner" SnapPlanner.cpp)
 target_link_libraries("${PROJECT_NAME}_planner"
   "${PROJECT_NAME}_constraint"
   "${PROJECT_NAME}_distance"
-  "${PROJECT_NAME}_path"
+  "${PROJECT_NAME}_trajectory"
   "${PROJECT_NAME}_statespace"
   "${PROJECT_NAME}_util")
