+++ aikido/src/planner/parabolic/CMakeLists.txt
@@ -5,6 +5,6 @@ aikido_add_library("${PROJECT_NAME}_planner_parabolic"
   ParabolicTimer.cpp)
 target_link_libraries("${PROJECT_NAME}_planner_parabolic"
   "${PROJECT_NAME}_hauserparabolicsmoother"
-  "${PROJECT_NAME}_path"
+  "${PROJECT_NAME}_trajectory"
   "${PROJECT_NAME}_util"
   "${PROJECT_NAME}_statespace")
