+++ aikido/tests/CMakeLists.txt
@@ -8,7 +8,7 @@ include_directories("${CMAKE_CURRENT_SOURCE_DIR}")
 add_subdirectory("constraint")
 add_subdirectory("distance")
 add_subdirectory("ompl")
-add_subdirectory("path")
 add_subdirectory("planner")
 add_subdirectory("statespace")
+add_subdirectory("trajectory")
 add_subdirectory("util")
