+++ aikido/tests/planner/CMakeLists.txt
@@ -4,5 +4,5 @@ aikido_add_test(test_SnapPlanner test_SnapPlanner.cpp)
 target_link_libraries(test_SnapPlanner
   "${PROJECT_NAME}_constraint"
   "${PROJECT_NAME}_distance"
-  "${PROJECT_NAME}_path"
+  "${PROJECT_NAME}_trajectory"
   "${PROJECT_NAME}_planner")
